vim: try to get OmniSharp running...
...and it does but it won't parse new csproj files with almost nothing
in them. Had to cobble stuff together from various StackOverflow and
GitHub posts to even get it running. Decided to try and build it to
maybe debug what's happening. Oh man!

- Have to LD_PRELOAD libcurl-compat or libgit2 will crash, which is used
  by libgit2sharp, whish is used by cake, a .NET build system I'd never
  heard of.

- Had to fix a bug where it bombs because Arch /etc/os_release does not
  contain a VERSION_ID.

- Now it's apperently tryong to install .NET Core 1.1 (still talking
  about the freaking _build syystem) here, and that doesn't work because
  it 404s, but may also be caused by the LD_PRELOAD thing.

Long story short, I should probably not use Arch for this, but an Ubuntu
instead. But with all the /etc/os-release dependent stuff going on in
both OmnuSharp and .NET Core, I'm not going to believe it when I gear
stuff about this or that company loving "Linux". You mean Ubuntu, RedHat
and Alpine!